+++ SisVenta/src/main/java/pe/edu/upeu/sisventa/repositorio/CompCarritoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upeu.sisventa.modelo.CompCarrito;
 
 @Repository
-public interface CompCarritoRepository extends JpaRepository<CompCarrito, Integer> {
+public interface CompCarritoRepository extends JpaRepository<CompCarrito, Long> {
 }
