+++ SisVenta/src/main/java/pe/edu/upeu/sisventa/repositorio/EmisorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upeu.sisventa.modelo.Emisor;
 
 @Repository
-public interface EmisorRepository extends JpaRepository<Emisor, Integer> {
+public interface EmisorRepository extends JpaRepository<Emisor, Long> {
 }
