+++ SisVenta/src/main/java/pe/edu/upeu/sisventa/repositorio/VentaDetalleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upeu.sisventa.modelo.VentaDetalle;
 
 @Repository
-public interface VentaDetalleRepository extends JpaRepository<VentaDetalle, Integer> {
+public interface VentaDetalleRepository extends JpaRepository<VentaDetalle, Long> {
 }
